Move script constructors  into a separate module
