+++ programs/budget_api/src/budget_instruction.rs
@@ -1,13 +1,8 @@
-use crate::budget_expr::{BudgetExpr, Condition};
-use crate::budget_state::BudgetState;
+use crate::budget_expr::BudgetExpr;
 use crate::id;
-use bincode::serialized_size;
 use chrono::prelude::{DateTime, Utc};
 use serde_derive::{Deserialize, Serialize};
 use solana_sdk::pubkey::Pubkey;
-use solana_sdk::script::Script;
-use solana_sdk::signature::{Keypair, KeypairUtil};
-use solana_sdk::system_instruction::SystemInstruction;
 use solana_sdk::transaction::Instruction;
 
 /// A smart contract.
@@ -40,87 +35,6 @@ impl BudgetInstruction {
         }
         keys.push((*contract, false));
         Instruction::new(id(), &BudgetInstruction::InitializeAccount(expr), keys)
-    }
-
-    pub fn new_initialize_account_script(
-        from: &Pubkey,
-        contract: &Pubkey,
-        lamports: u64,
-        expr: BudgetExpr,
-    ) -> Script {
-        let space = serialized_size(&BudgetState::new(expr.clone())).unwrap();
-        Script::new(vec![
-            SystemInstruction::new_program_account(&from, contract, lamports, space, &id()),
-            Self::new_initialize_account(contract, expr),
-        ])
-    }
-
-    /// Create a new payment script.
-    pub fn new_payment_script(from: &Pubkey, to: &Pubkey, lamports: u64) -> Script {
-        let contract = Keypair::new().pubkey();
-        let expr = BudgetExpr::new_payment(lamports, to);
-        Self::new_initialize_account_script(from, &contract, lamports, expr)
-    }
-
-    /// Create a postdated payment script.
-    pub fn new_on_date_script(
-        from: &Pubkey,
-        to: &Pubkey,
-        contract: &Pubkey,
-        dt: DateTime<Utc>,
-        dt_pubkey: &Pubkey,
-        cancelable: Option<Pubkey>,
-        lamports: u64,
-    ) -> Script {
-        let expr = if let Some(from) = cancelable {
-            BudgetExpr::Or(
-                (
-                    Condition::Timestamp(dt, *dt_pubkey),
-                    Box::new(BudgetExpr::new_payment(lamports, to)),
-                ),
-                (
-                    Condition::Signature(from),
-                    Box::new(BudgetExpr::new_payment(lamports, &from)),
-                ),
-            )
-        } else {
-            BudgetExpr::After(
-                Condition::Timestamp(dt, *dt_pubkey),
-                Box::new(BudgetExpr::new_payment(lamports, to)),
-            )
-        };
-
-        Self::new_initialize_account_script(from, contract, lamports, expr)
-    }
-
-    /// Create a multisig payment script.
-    pub fn new_when_signed_script(
-        from: &Pubkey,
-        to: &Pubkey,
-        contract: &Pubkey,
-        witness: &Pubkey,
-        cancelable: Option<Pubkey>,
-        lamports: u64,
-    ) -> Script {
-        let expr = if let Some(from) = cancelable {
-            BudgetExpr::Or(
-                (
-                    Condition::Signature(*witness),
-                    Box::new(BudgetExpr::new_payment(lamports, to)),
-                ),
-                (
-                    Condition::Signature(from),
-                    Box::new(BudgetExpr::new_payment(lamports, &from)),
-                ),
-            )
-        } else {
-            BudgetExpr::After(
-                Condition::Signature(*witness),
-                Box::new(BudgetExpr::new_payment(lamports, to)),
-            )
-        };
-
-        Self::new_initialize_account_script(from, contract, lamports, expr)
     }
 
     pub fn new_apply_timestamp(
