+++ programs/budget_api/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod budget_expr;
 pub mod budget_instruction;
+pub mod budget_script;
 pub mod budget_state;
 pub mod budget_transaction;
 pub mod payment_plan;
